+++ src/components/task.js
@@ -47,12 +47,4 @@ const createTask = () => {
   );
 };
 
-const createTasksList = (count) => {
-  let tasksList = ``;
-  for (let i = 0; i < count; i++) {
-    tasksList += createTask();
-  }
-  return tasksList;
-};
-
-export default createTasksList;
+export default createTask;
